♻️  [Refactor]: Remove useless layout component
기존에 MyBookmark.template.tsx가 하우스, 라운지, 게시글 템플릿을 호출하는 중첩 구조로 되어 있어 혼란을 줄 수 있어 계층 구조를 명확히 하기 위해 MyBookmark 페이지에서 템플릿을 직접 관리하도록 수정.
